+++ R/misc.R
@@ -8,7 +8,10 @@
 #'
 #' @return A data frame or matrix with the new column names.
 #' @export
-setcolnames <- function(object = nm, nm) {
+setcolnames <- function(object, nm) {
+  if (length(nm) != ncol(object)) {
+    stop("Length of 'nm' must equal the number of columns of 'object'")
+  }
   colnames(object) <- nm
   object
 }
@@ -23,7 +26,10 @@ setcolnames <- function(object = nm, nm) {
 #'
 #' @return A data frame or matrix with the new row names.
 #' @export
-setrownames <- function(object = nm, nm) {
+setrownames <- function(object, nm) {
+  if (length(nm) != nrow(object)) {
+    stop("Length of 'nm' must equal the number of rows of 'object'")
+  }
   rownames(object) <- nm
   object
 }
@@ -40,34 +46,48 @@ setrownames <- function(object = nm, nm) {
 #' @return A long data.table with two columns: 'name' and 'value'.
 #' @importFrom data.table data.table
 #' @export
+#' @examples
+#' library(easybio)
+#' list2dt(list(a = c(1, 1), b = c(2, 2)))
 list2dt <- function(x) {
   data.table(name = rep(names(x), sapply(x, length)), value = unlist(x))
 }
 
 
-#' Split a Matrix into Smaller Submatrices by Column
+#' Split a Matrix into Smaller Sub-matrices by Column or Row
 #'
-#' This function splits a matrix into multiple smaller matrices by column.
+#' This function splits a matrix into multiple smaller matrices by column or row.
 #' It is useful for processing large matrices in chunks, such as when performing
 #' analysis on a single computer with limited memory.
 #'
 #' @param matrix A numeric or logical matrix to be split.
-#' @param chunk_size The number of columns to include in each smaller matrix.
+#' @param chunk_size The number of columns or rows to include in each smaller matrix.
+#' @param column  Divided by column(default is `TRUE`)
 #'
-#' @return A list of smaller matrices, each with `chunk_size` columns.
+#' @return A list of smaller matrices, each with `chunk_size` columns or rows.
 #' @export
-split_matrix <- function(matrix, chunk_size) {
-  chunk_number <- ifelse(ncol(matrix) %% chunk_size == 0,
-    ncol(matrix) / chunk_size - 1,
-    floor(ncol(matrix) / chunk_size)
+#' @examples
+#' library(easybio)
+#' split_matrix(mtcars, chunk_size = 2)
+#' split_matrix(mtcars, chunk_size = 5, column = FALSE)
+split_matrix <- function(matrix, chunk_size, column = TRUE) {
+  n <- ifelse(column, ncol(matrix), nrow(matrix))
+  chunk_number <- ifelse(n %% chunk_size == 0,
+    n / chunk_size - 1,
+    floor(n / chunk_size)
   )
-  message(sprintf("matrix was divided to %f chunks", chunk_number + 1))
+  message(sprintf("matrix was divided to %d chunks", chunk_number + 1))
   start_end <- lapply(0:chunk_number, function(x) {
     c(1, chunk_size) + (chunk_size * x)
   })
-  start_end[[chunk_number + 1]][[2]] <- ncol(matrix)
-  start_end
-  matrix_divided <- lapply(start_end, function(x) matrix[, x[[1]]:x[[2]]])
+  start_end[[chunk_number + 1]][[2]] <- n
+  matrix_divided <- lapply(start_end, function(x) {
+    if (column) {
+      matrix[, x[[1]]:x[[2]], drop = FALSE]
+    } else {
+      matrix[x[[1]]:x[[2]], , drop = FALSE]
+    }
+  })
 
   matrix_divided
 }
@@ -96,41 +116,36 @@ get_attr <- function(x, attr_name) {
 #' @param nodes A named list where each element is a vector.
 #'
 #' @return A data.table representing the graph, with columns for the node names
-#'   (`node_x` and `node_y`) and the weight of the edge (`weight`).
+#'   (`node_1` and `node_2`) and the weight of the edge (`interWeight`).
 #' @import data.table
 #' @export
 list2graph <- function(nodes) {
-  node_x <- c()
-  node_y <- c()
-  weight <- numeric()
-  for (i in seq_along(nodes)) {
-    j <- i + 1
-    while (j <= length(nodes)) {
-      node_x <- append(node_x, names(nodes[i]))
-      node_y <- append(node_y, names(nodes[j]))
-      weight <- append(weight, intersect(nodes[[i]], nodes[[i]]) |> length())
+  comb2 <- combn(names(nodes), m = 2, simplify = FALSE)
+  inter <- lapply(comb2, \(x) length(intersect(nodes[[x[[1]]]], nodes[[x[[2]]]])))
 
-      j <- j + 1
-    }
-  }
-
-  net <- data.table(node_x = node_x, node_y = node_y, weight = weight)
-  net
+  data.table(
+    node1 = sapply(comb2, \(x) x[[1]]),
+    node2 = sapply(comb2, \(x) x[[2]]),
+    interWeight = as.integer(inter)
+  )
 }
 
 
-#' Perform Summary Analysis by Group Using an Index
+#' Perform Summary Analysis by Group Using an column Index
 #'
-#' This function applies a specified function to each group defined by an index,
+#' This function applies a specified function to each group defined by an column index,
 #' and returns a summary of the results. It is useful for summarizing data by
-#' group when the groups are defined by an index rather than a named column.
+#' group when the groups are defined by an  column index.
 #'
 #' @param f A function that takes a single argument and returns a summary of the data.
 #' @param x A data frame or matrix containing the data to be summarized.
-#' @param idx A vector of indices or group names that define the groups.
+#' @param idx A list of indices or group names that define the column groups.
 #'
-#' @return A data frame or matrix containing the summary statistics for each group.
+#' @return A list containing the summary statistics for each group.
 #' @export
+#' @examples
+#' library(easybio)
+#' groupStatI(f = \(x) x + 1, x = mtcars, idx = list(c(1, 10), 2))
 groupStatI <- function(f, x, idx) {
   sapply(idx, \(.x) force(f)(x[.x]), simplify = FALSE)
 }
@@ -144,11 +159,14 @@ groupStatI <- function(f, x, idx) {
 #' @param f A function that takes a single argument and returns a summary of the data.
 #' @param x A data frame or matrix containing the data to be summarized.
 #' @param xname A character vector containing the names of the variables in `x`.
-#' @param patterns A character vector of regular expressions that define the groups.
+#' @param patterns A list of regular expressions that define the groups.
 #'
-#' @return A data frame or matrix containing the summary statistics for each group.
+#' @return A list containing the summary statistics for each group.
 #' @export
-groupStat <- function(f, x, xname = names(x), patterns) {
+#' @examples
+#' library(easybio)
+#' groupStat(f = \(x) x + 1, x = mtcars, patterns = list("mp", "t"))
+groupStat <- function(f, x, xname = colnames(x), patterns) {
   idx <- lapply(patterns, \(.x) which(xname %like% .x))
   groupStatI(f, x, idx)
 }